refactoring: reduce duplication, make things more readable
